32bit fix

+++ src/baldr/location.cc
@@ -148,7 +148,7 @@ Location Location::FromRapidJson(const rapidjson::Value& d, unsigned int default
   location.way_id_ = GetOptionalFromRapidJson<uint64_t>(d, "/way_id");
 
   location.minimum_reachability_ = GetFromRapidJson<unsigned int>(d, "/minimum_reachability", default_reachability);
-  location.radius_ = GetFromRapidJson<unsigned long>(d, "/radius", default_radius);
+  location.radius_ = GetFromRapidJson<unsigned int>(d, "/radius", default_radius);
 
   return location;
 }
